refactor: ModelBubble import paths and create ModelBubble component for model display
